add:コンポーネントに切り出して管理しやすくする実装
